Revert "move fullscreen into acuantcapturecanvas"
This reverts commit ffd83e51d3abe450fd997cae91fa20bdf4e86703.
